# apps/api/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Nimbus Platform API
  description: |
    AI Native Internal Developer Platform REST API (핵심 엔드포인트 스냅샷).
    상세 시나리오 스펙은 docs/api/API-*.md 를 함께 본다.
    저장소 소스: docs/api/openapi.yaml · 런타임 서빙: classpath:openapi.yaml (동기화 필수).
  version: 0.5.0
  contact:
    name: Nimbus Platform
servers:
  - url: http://localhost:8080
    description: Local
...
paths:
...
  /api/v1/services:
    get:
      tags: [Services]
      summary: List services
      security:
        - bearerAuth: []
      parameters:
        - name: workspaceId
          in: query
          schema: { type: string, format: uuid }
        - name: projectId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Service list
  /api/v1/services/{serviceId}:
    get:
      tags: [Services]
      summary: Get service
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/serviceId"
      responses:
        "200":
          description: Service
...
  /api/v1/notifications:
    get:
      tags: [Auth]
      summary: List in-app notifications
      security:
        - bearerAuth: []
      parameters:
        - name: workspaceId
          in: query
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Notifications
  /api/v1/notifications/unread-count:
    get:
      tags: [Auth]
      summary: Unread notification count
      security:
        - bearerAuth: []
      parameters:
        - name: workspaceId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Count
  /api/v1/notifications/sync:
    post:
      tags: [Auth]
      summary: Sync notifications from incidents/sagas/pipelines
      security:
        - bearerAuth: []
      parameters:
        - name: workspaceId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Sync result
  /api/v1/services/{serviceId}/tags:
    put:
      tags: [Services]
      summary: Replace service tags
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/serviceId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                tags:
                  type: array
                  items: { type: string }
      responses:
        "200":
          description: Updated service
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    serviceId:
      name: serviceId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
security:
  - bearerAuth: []